Moves check_instance to htypes
To centralize this and make it clear we have two ways of checking types.
I don't want to merge them because the input type checker is less strict.
It doesn't recursively go check the values of inputs. So there's a TODO:
to figure out what to do, and how to unify. In the meantime they're in
the same file.
