+++ hamilton/lifecycle/default.py
@@ -4,15 +4,10 @@ import logging
 import pdb
 import pprint
 import random
-import sys
 import time
 from typing import Any, Callable, Dict, List, Optional, Union
 
-if sys.version_info >= (3, 9):
-    from typing import Literal
-else:
-    Literal = None
-
+from hamilton import htypes
 from hamilton.lifecycle import NodeExecutionHook
 from hamilton.lifecycle.api import NodeExecutionMethod
 
@@ -348,64 +343,6 @@ class SlowDownYouMoveTooFast(NodeExecutionHook):
         pass
 
 
-def check_instance(obj: Any, type_: Any) -> bool:
-    """This function checks if an object is an instance of a given type. It supports generic types as well.
-
-    :param obj: The object to check.
-    :param type_: The type to check against. This can be a generic type like List[int] or Dict[str, Any].
-    :return: True if the object is an instance of the type, False otherwise.
-    """
-    if type_ == Any:
-        return True
-    # Get the origin of the type (i.e., the base class for generic types)
-    origin = getattr(type_, "__origin__", None)
-
-    # If the type has an origin, it's a generic type
-    if origin is not None:
-        # If the type is a Union type
-        if origin is Union:
-            return any(check_instance(obj, t) for t in type_.__args__)
-        elif origin is Literal:
-            return obj in type_.__args__
-        # Check if the object is an instance of the origin of the type
-        elif not isinstance(obj, origin):
-            return False
-
-        # If the type has arguments (i.e., it's a parameterized generic type like List[int])
-        if hasattr(type_, "__args__"):
-            # Get the element type(s) of the generic type
-            element_type = type_.__args__
-
-            # If the object is a dictionary
-            if isinstance(obj, dict):
-                all_items_meet_condition = True
-
-                # Iterate over each key-value pair in the dictionary
-                for key, value in obj.items():
-                    # Check if the key is an instance of the first element type and the value is an instance of the second element type
-                    key_is_correct_type = check_instance(key, element_type[0])
-                    value_is_correct_type = check_instance(value, element_type[1])
-
-                    # If either the key or the value is not the correct type, set the flag to False and break the loop
-                    if not key_is_correct_type or not value_is_correct_type:
-                        all_items_meet_condition = False
-                        break
-
-                # Return the result
-                return all_items_meet_condition
-
-            # If the object is a list, set, or tuple
-            elif isinstance(obj, (list, set, tuple)):
-                element_type = element_type[0]
-                for i in obj:
-                    if not check_instance(i, element_type):
-                        return False
-                return True
-
-    # If the type is not a generic type, just use isinstance
-    return isinstance(obj, type_)
-
-
 class FunctionInputOutputTypeChecker(NodeExecutionHook):
     """This lifecycle hook checks the input and output types of a function.
 
@@ -436,7 +373,7 @@ class FunctionInputOutputTypeChecker(NodeExecutionHook):
         """Checks that the result type matches the expected node return type."""
         if self.check_input:
             for input_name, input_value in node_kwargs.items():
-                if not check_instance(input_value, node_input_types[input_name]):
+                if not htypes.check_instance(input_value, node_input_types[input_name]):
                     raise TypeError(
                         f"Node {node_name} received an input of type {type(input_value)} for {input_name}, expected {node_input_types[input_name]}"
                     )
@@ -457,7 +394,7 @@ class FunctionInputOutputTypeChecker(NodeExecutionHook):
         """Checks that the result type matches the expected node return type."""
         if self.check_output:
             # Replace the isinstance check in your code with check_instance
-            if not check_instance(result, node_return_type):
+            if not htypes.check_instance(result, node_return_type):
                 raise TypeError(
                     f"Node {node_name} returned a result of type {type(result)}, expected {node_return_type}"
                 )
